% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_formatted_gapit.R
\name{get_formatted_gapit}
\alias{get_formatted_gapit}
\title{Import and format GAPIT output results for \code{\link[=plot_manhattan]{plot_manhattan()}} usage}
\usage{
get_formatted_gapit(result_directory, gapit_cutOff = 0.05, verbose = TRUE)
}
\arguments{
\item{result_directory}{: str     Directory GAPIT package exported one folder
for each trait.}

\item{gapit_cutOff}{: double  Value used for 'cutOff' argument during
GAPIT analysis.}

\item{verbose}{: bool    Print which trait is being formatted if TRUE}
}
\value{
GAPIT.thresh object.
}
\description{
Import and format GAPIT output results for \code{\link[=plot_manhattan]{plot_manhattan()}} usage
}
\examples{
\dontrun{
get_formatted_gapit(result_directory = "./GAPIT")
}

}
\seealso{
\itemize{
\item \code{\link[=extract_markers]{extract_markers()}} for significant marker extraction.
\item \code{\link[=plot_manhattan]{plot_manhattan()}} for beautified Manhattan plot.
}
}
