% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manhattan.R
\name{plot_manhattan}
\alias{plot_manhattan}
\title{Manhattan plot beautifier adapted from GWASpoly::manhattan.plot customized
and optimized for data obtained with \code{\link[GWASpoly:set.threshold]{GWASpoly::set.threshold()}} or
\code{\link[=get_formatted_gapit]{get_formatted_gapit()}} command.}
\usage{
plot_manhattan(
  data,
  traits = NULL,
  models = NULL,
  chrom = NULL,
  chrom_color = c("#219ebc", "#8ecae6"),
  significant_color = "#fb8500",
  point_size = 1.5,
  point_alpha = NA,
  threshold_line_color = "grey50",
  threshold_line_type = 2,
  significant_markers = NULL,
  gap_size = 0.01
)
}
\arguments{
\item{data, traits, models, chrom}{See GWASpoly::manhattan.plot documentation}

\item{chrom_color}{: vec(str, str)  List of 2 string color for
alternate coloration of marker.}

\item{significant_color}{: str            String color for significant
marker, NULL  leaves significant
markers with the same color as
not significant ones.}

\item{point_size}{: float          Diameter of the points.}

\item{point_alpha}{: float          Points transparency.}

\item{threshold_line_color}{: str            Threshold line color.}

\item{threshold_line_type}{: int            Threshold line type.}

\item{significant_markers}{: data.frame     Dataframe generated by the
function \code{\link[=extract_markers]{extract_markers()}}
containing only the significant
marker to highlight with
'significant_color'. NULL value
will highlight all markers
according to 'significant_color'
parameter.}

\item{gap_size}{: float          Total size multiplier to define
gap size between chromosomes.}
}
\value{
ggplot2::ggplot object.
}
\description{
Manhattan plot beautifier adapted from GWASpoly::manhattan.plot customized
and optimized for data obtained with \code{\link[GWASpoly:set.threshold]{GWASpoly::set.threshold()}} or
\code{\link[=get_formatted_gapit]{get_formatted_gapit()}} command.
}
\examples{
\dontrun{
plot_manhattan(data = data_with_threshold, significant_color = NULL)
}

}
\seealso{
\itemize{
\item \code{\link[=extract_markers]{extract_markers()}} for significant marker extraction.
\item \code{\link[=save_manhattan]{save_manhattan()}} for pre-formatted plot saving options.
}
}
