% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_manhattan.R
\name{save_manhattan}
\alias{save_manhattan}
\title{Save beautified Manhattan plot with the right ratio}
\usage{
save_manhattan(gwas_plot, file_name)
}
\arguments{
\item{gwas_plot}{: ggplot  ggplot2::ggplot object generated by
'GWASpoly_beautifier::plot_manhattan'.}

\item{file_name}{: str     File name to save Manhattan plot.}
}
\description{
Save beautified Manhattan plot with the right ratio
}
\examples{
\dontrun{
save_manhattan(gwas_plot = p, file_name = "vine_maturity_manhattan.png")
}
}
