% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_markers.R
\name{extract_markers}
\alias{extract_markers}
\title{Extract all/significant markers from GWASpoly.thresh data obtained with
GWASpoly::set.threshold(...) command.}
\usage{
extract_markers(data, significant_only = TRUE)
}
\arguments{
\item{data}{: GWASpoly.thresh  GWASpoly threshold data.}

\item{significant_only}{: bool             Extract only markers above
threshold.}
}
\value{
data.frame object with all requested markers.
}
\description{
Extract all/significant markers from GWASpoly.thresh data obtained with
GWASpoly::set.threshold(...) command.
}
\examples{
\dontrun{
significant_markers <- extract_markers(data = data_with_threshold)
}
}
